+++ packages/dashboard-backend/src/cheServerApiProxy.ts
@@ -9,25 +9,11 @@
  * Contributors:
  *   Red Hat, Inc. - initial API and implementation
  */
-import { FastifyInstance, FastifyReply, FastifyRequest, RouteShorthandOptions } from 'fastify';
+import { FastifyInstance } from 'fastify';
 import fastifyHttpProxy from 'fastify-http-proxy';
-import { ReplyDefault } from 'fastify/types/utils';
 
 export function registerCheServerApiProxy(server: FastifyInstance, cheApiUpstream: string, origin: string) {
   console.log(`Che Server API proxy is running and proxying to "${cheApiUpstream}".`);
-
-  // fake JSON RPC for Che websocket API
-  // because the real proxy fails to some reason
-  // but since che workspace and devworkspace are not expected to work at the same time
-  // faking is an easier solution
-  server.get(`/api/websocket`, { websocket: true } as RouteShorthandOptions, (connection: FastifyRequest) => {
-    connection.socket.on('message', message => {
-      const data = JSON.parse(message);
-      if (data!.id && data!.jsonrpc) {
-        (connection.socket as any).send(JSON.stringify({ jsonrpc: data.jsonrpc, id: data.id, result: [] }));
-      }
-    });
-  });
 
   server.register(fastifyHttpProxy, {
     upstream: cheApiUpstream,
@@ -40,16 +26,6 @@ export function registerCheServerApiProxy(server: FastifyInstance, cheApiUpstrea
         return Object.assign({ ...headers }, { origin });
       }
     }
-  });
-
-  // OPTIONS request to '/api/' fails when running the backend locally. This hook returns stub object to prevent such misbehavior.
-  server.addHook('onRequest', (request, reply, done) => {
-    if ((request.url === '/api' || request.url === '/api/') && request.method === 'OPTIONS') {
-      return reply.send({
-        implementationVersion: 'Local Run',
-      });
-    }
-    done();
   });
 
   server.register(fastifyHttpProxy, {
